src/Python/sentimant_analysisi/pyproject.toml
[project]
name = "sentimant-analysis"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "aren", email = "code@arencore.me" }
]
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn>=0.30.1",
    "python-dotenv>=1.0.1",
    "sqlalchemy>=2.0.31",
    "psycopg2-binary>=2.9.9",
    "openai>=1.35.13",
    "requests>=2.32.3",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/sentimant_analysis"]
